Update TeamController.php
The switchCurrentTeam function is sent the ID of a team and sends that ID to the $request->user()->switchToTeam($team); function. However, this function expects an object of team properties. The result is an error when you try to switch teams. Adding in the line to find the team from the ID solves this problem.