Validation Fails if no stripe_token is provided
If Spark is setup to have "CollectBillingAddress" enabled and the user trying to subscribe has already inserted a payment method, Spark tries to validate the billing address.
If the user has selected "use existing payment method" in the subscribe form, the validation fails because the stripe_token is not sent to the backend (it's not required). Since we mark it as as not required in the `mergeCardCountryIntoRequest` method, it shouldn't be required during the validation for this reason too.