Update UpdateContactInformationOnStripe.php
When Team Billing is being used, the 'stripe_id' database field stays empty for the User Model. When User changes their email address, the listener will not work for changing the email address for the corresponding customer in Stripe, since their is no stripe_id stored in the database. This updated listener will look for the stripe_id at the currentTeam, in order to find the matching customer in Stripe, so that email addresses can be updated.