[8.0] update webpack stub
I'm not sure what the intent of including the RTL css file in here was. Maybe an older version of Mix behaved differently?

Whether I compile with or without this code, I get the same output.  According to the [documentation](https://laravel.com/docs/5.8/mix#versioning-and-cache-busting):

> The version method will automatically append a unique hash to the filenames of all compiled files, allowing for more convenient cache busting

Since the RTL file is compiled, it is automatically versioned.