Update webpack.mix.js
only create RTL css and version files when running in production

in local development, there is a lot of iteration which leads to a lot of code compilation. this can often be a timely operation, which can be a point of frustration.

by moving this code to run only in production, I have seen reductions of 5 to 10 seconds in compilations.