Update notifications.blade.php
if a user has upgraded to FontAwesome 5, the notifications break. due to something with how FA5 replaces the `<i>` tags with `<svg>`s and the Vue conditional, you end up with multiple SVGs and the notifications don't load correctly.

wrapping the icon and the words in a `<div>` fixes it, although I'm not exactly sure why.