Change disk storage to default
Hard-coding public into the interaction means that this doesn't follow convention with the rest of laravel, e.g. ENV settings. We just ran into an issue with a load-balanced environment because this was hard-coded to public, rather than the cloud / S3 instance we were using for file storage. 

Think it's a good option to allow people to control this. 🤷‍♂️